// types/user_types.d.ts
export type userStats = {
  user_id: string;
  created_at: string;
  last_daily_tab: string;
  opening_ok: number;
  ending_ok: number;
  description_ok: number;
  image_ok: number;
  animedle_ok: number;
  all_ok: number;
  user_name: string;
  current_stage: number;
  daily_complete: boolean;
  game1_status: number;
  game2_status: number;
  game3_status: number;
  game4_status: number;
  game5_status: number;
};